fix string formatting bug

+++ src/jni/ObjectModel/Tool.h
@@ -130,7 +130,7 @@ namespace OM
 	bool UpdateToolSpindle(const size_t toolIndex, const int8_t spindleIndex);
 	bool UpdateToolSpindleRpm(const size_t toolIndex, const int32_t rpm);
 
-	void SetCurrentTool(const size_t toolIndex);
+	void SetCurrentTool(const int32_t toolIndex);
 	Tool* GetCurrentTool();
 }
